% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binCLT.R
\name{binCLT}
\alias{binCLT}
\title{binCLT}
\usage{
binCLT(n, iter, p = 0.5, ...)
}
\arguments{
\item{n}{sample size}

\item{iter}{number of iterations}

\item{p}{the probability of success in each binomial trial}

\item{...}{additional parameters}
}
\value{
A histogram of sample means and a CLT-based normal curve.
}
\description{
binCLT
}
\examples{
binCLT(n = 30, iter = 100, p = 0.4)
}
