% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Linearmodel.R
\name{lab3}
\alias{lab3}
\title{Linear Model Calculation}
\usage{
lab3(x)
}
\arguments{
\item{x}{}
}
\value{
A linear regression model.
}
\description{
Calculate a linear regression model using the 'lm' function.
}
\examples{
(lab3(1:10))
}
