% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Task8Lab4.R
\name{lab4}
\alias{lab4}
\title{Task 8 in lab 4}
\usage{
lab4(x)
}
\arguments{
\item{x}{A quantitative vector}
}
\value{
A vector of squared components
}
\description{
Task 8 in lab 4
}
\examples{
Task8Lab4(1:10)
}
