% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logbin.R
\name{logbin}
\alias{logbin}
\title{logbin}
\usage{
logbin(x = c(3, 3, 4, 3, 4, 5, 5, 4), param = seq(0, 1, length = 1000))
}
\arguments{
\item{x}{vector}

\item{param}{the probability of success in a single trial.}
}
\description{
natural logarithm of binomial probabilities
}
\examples{
logbin(x=c(9,9,1,9,9,9),param=seq(0,1,length=1000))
}
