% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.R
\name{mybin}
\alias{mybin}
\title{mybin}
\usage{
mybin(iter = 100, n = 10, p = 0.7)
}
\arguments{
\item{iter}{number of iterations for the binomial simulation.}

\item{n}{sample size for each trial}

\item{p}{probability of success in each binomial trial}
}
\value{
A barplot of the proportions of successes and a table of success counts.
}
\description{
perform a binomial simulation and create a barplot of the proportions of successes
}
\examples{
mybin(iter = 100, n = 20, p = 0.6)
}
