% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{myboot2}
\usage{
myboot2(x, fun = "mean", iter = 10000, alpha = 0.05, cx = 1.5, xlab = fun, ...)
}
\arguments{
\item{x}{the data}

\item{fun}{function using/the value we want to calculate}

\item{iter}{number of iterations}

\item{alpha}{bootstrap interval}

\item{cx}{size of the text in the graph}

\item{xlab}{label for the x-axis on the histogram}

\item{...}{additional parameters}
}
\value{
A list containing the calculated bootstrap interval, the function used, the input data, and the vector of bootstrap statistics.
}
\description{
bootstrap analysis on a given dataset
}
\examples{
myboot2(iter = 10000, x=rnorm(50), fun = "mean", alpha = 0.05, xlab = "mean")
}
