% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfunction.R
\name{myfunction}
\alias{myfunction}
\title{My first function}
\usage{
myfunction(x)
}
\arguments{
\item{x}{A quantitative vector}
}
\value{
A vector of squared components
}
\description{
My first function
}
\examples{
myfunction(1:10)
}
