% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlik.R
\name{mymaxlik}
\alias{mymaxlik}
\title{mymaxlik}
\usage{
mymaxlik(lfun, x, param, ...)
}
\arguments{
\item{lfun}{the likelihood function}

\item{x}{vector of the observe data}

\item{param}{A numeric vector of parameter values over which the likelihood
function will be evaluated}

\item{...}{additional parameters}
}
\value{
list relate to the MLE
}
\description{
maximum likelihood function for repeated sampling from same distribution
}
\examples{
mymaxlik(x=c(9,9,1,9,9,9),param=seq(0,1,length=1000),
          lfun=logbin,xlab=expression(pi),main="Binomial",cex.main=2)
}
