% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{myncurve}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{mean (mu) of the normal distribution.}

\item{sigma}{standard deviation (sigma) of the normal distribution}

\item{a}{a point up to which you want to calculate the area under the curve.}
}
\value{
the area (probability, P(X<=a)) of the normal distribution
}
\description{
myncurve
}
\examples{
myncurve(4,4,5)
}
