% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{ntickets}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{number of seats on the flight}

\item{gamma}{probability of overbooking}

\item{p}{probability of a passenger will show}
}
\value{
number of seats sold(nd is calculated by the discrete distribution and nc is calculated by the discrete distribution )
}
\description{
Trying to find the optimization number of ticket to sell to avoid overbooking problem
}
\examples{
ntickets(200, 0.02, 0.95)
}
