% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterhist.R
\name{scatterhist}
\alias{scatterhist}
\title{Scatter histogrm}
\usage{
scatterhist(x, y, xlab = "", ylab = "")
}
\arguments{
\item{x}{X variable as a vector}

\item{y}{Y variable as a vector}

\item{xlab}{X label}

\item{ylab}{Y label}
}
\value{
A sophisticated plot
}
\description{
Scatter histogrm
}
\examples{
\dontrun{with(ddt, scatterhist(LENGTH,WEIGHT, xlab="LENGTH"))}
}
